Move test factory to tests utilities
